refactor: 명확하게 이름 변경
