+++ src/components/RowMovie/RowMovie.css
@@ -5,7 +5,7 @@
   font-family: 'SF Pro Display', sans-serif;
 }
 
-.row-title {
+.row_title {
   font-size: 20px;
   font-weight: 700;
   letter-spacing: 0.76px;
@@ -14,7 +14,7 @@
   padding-left: 4px;
 }
 
-.row-posters {
+.row_posters {
   display: flex;
   overflow-x: scroll;
   overflow-y: hidden;
@@ -24,13 +24,13 @@
   scroll-behavior: smooth;
 }
 
-.row-posters::-webkit-scrollbar {
+.row_posters::-webkit-scrollbar {
   width: 0;
   height: 0;
   display: none;
 }
 
-.row-poster {
+.row_poster {
   width: 103px;
   height: 161px;
   object-fit: cover;
@@ -39,11 +39,11 @@
   flex-shrink: 0;
 }
 
-.row-poster:hover {
+.row_poster:hover {
   transform: scale(1.05);
 }
 
-.row-poster-large {
+.row_poster-large {
   width: 160px;
   height: auto;
 }
